Stop redundant saving of user record, add error if the user is invalid.
This commit is useful for my own use, but not necessarily for general
consumption. The functionality of failing the session if the user is
invalid is something that should probably be configurable. Nonetheless,
it seems odd to allow session creation with an invalid new user.

As a side note, I've removed the call to save because assigning
attempted_record seems to ensure that authlogic will itself save the
record.